Fix user related issue

+++ src/main/java/com/services/todoservice/dto/UsersDTO.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 public class UsersDTO {
-    int userID;
+    Long userID;
     String username;
     String password;
 }
